requirements.txt
Adafruit-Blinka==8.20.1
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-motor==3.4.11
adafruit-circuitpython-pca9685==3.4.10
adafruit-circuitpython-register==1.9.16
adafruit-circuitpython-requests==2.0.1
adafruit-circuitpython-servokit==1.3.15
adafruit-circuitpython-typing==1.9.4
Adafruit-PlatformDetect==3.49.0
Adafruit-PureIO==1.1.11
colorzero==2.0
gpiozero==2.0
numpy==1.25.2
opencv-python==4.5.3.56
packaging==23.2
picamera==1.13
pigpio==1.78
Pillow==10.0.1
pkg_resources==0.0.0
pyftdi==0.55.0
pyserial==3.5
pytesseract==0.3.10
pyusb==1.2.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
sysv-ipc==1.1.0
tesseract==0.1.3
typing_extensions==4.7.1
